apps/cic-ussd/requirements.txt
alembic==1.4.2
bcrypt==3.2.0
celery==4.4.7
cic-eth[services]==0.12.2a3
cic-notify~=0.4.0a10
cic-types~=0.1.0a14
confini~=0.4.1a1
phonenumbers==8.12.12
psycopg2==2.8.6
python-i18n[YAML]==0.3.9
pyxdg==0.27
redis==3.5.3
semver==2.13.0
SQLAlchemy==1.3.20
tinydb==4.2.0
transitions==0.8.4
uWSGI==2.0.19.1